+++ celery-codegen/src/task.rs
@@ -24,6 +24,7 @@ enum TaskAttr {
     MaxRetries(syn::LitInt),
     MinRetryDelay(syn::LitInt),
     MaxRetryDelay(syn::LitInt),
+    RetryForUnexpected(syn::LitBool),
     AcksLate(syn::LitBool),
     Bind(syn::LitBool),
     OnFailure(syn::Ident),
@@ -41,6 +42,7 @@ struct Task {
     max_retries: Option<syn::LitInt>,
     min_retry_delay: Option<syn::LitInt>,
     max_retry_delay: Option<syn::LitInt>,
+    retry_for_unexpected: Option<syn::LitBool>,
     acks_late: Option<syn::LitBool>,
     original_args: Vec<syn::FnArg>,
     inputs: Option<Punctuated<FnArg, Comma>>,
@@ -123,6 +125,16 @@ impl TaskAttrs {
             .next()
     }
 
+    fn retry_for_unexpected(&self) -> Option<syn::LitBool> {
+        self.attrs
+            .iter()
+            .filter_map(|a| match a {
+                TaskAttr::RetryForUnexpected(r) => Some(r.clone()),
+                _ => None,
+            })
+            .next()
+    }
+
     fn acks_late(&self) -> Option<syn::LitBool> {
         self.attrs
             .iter()
@@ -181,6 +193,7 @@ mod kw {
     syn::custom_keyword!(max_retries);
     syn::custom_keyword!(min_retry_delay);
     syn::custom_keyword!(max_retry_delay);
+    syn::custom_keyword!(retry_for_unexpected);
     syn::custom_keyword!(acks_late);
     syn::custom_keyword!(bind);
     syn::custom_keyword!(on_failure);
@@ -218,6 +231,10 @@ impl parse::Parse for TaskAttr {
             input.parse::<kw::max_retry_delay>()?;
             input.parse::<Token![=]>()?;
             Ok(TaskAttr::MaxRetryDelay(input.parse()?))
+        } else if lookahead.peek(kw::retry_for_unexpected) {
+            input.parse::<kw::retry_for_unexpected>()?;
+            input.parse::<Token![=]>()?;
+            Ok(TaskAttr::RetryForUnexpected(input.parse()?))
         } else if lookahead.peek(kw::acks_late) {
             input.parse::<kw::acks_late>()?;
             input.parse::<Token![=]>()?;
@@ -252,6 +269,7 @@ impl Task {
             max_retries: attrs.max_retries(),
             min_retry_delay: attrs.min_retry_delay(),
             max_retry_delay: attrs.max_retry_delay(),
+            retry_for_unexpected: attrs.retry_for_unexpected(),
             acks_late: attrs.acks_late(),
             original_args: Vec::new(),
             inputs: None,
@@ -459,6 +477,11 @@ impl ToTokens for Task {
             .as_ref()
             .map(|r| quote! { Some(#r) })
             .unwrap_or_else(|| quote! { None });
+        let retry_for_unexpected = self
+            .retry_for_unexpected
+            .as_ref()
+            .map(|r| quote! { Some(#r) })
+            .unwrap_or_else(|| quote! { None });
         let acks_late = self
             .acks_late
             .as_ref()
@@ -570,6 +593,7 @@ impl ToTokens for Task {
                         max_retries: #max_retries,
                         min_retry_delay: #min_retry_delay,
                         max_retry_delay: #max_retry_delay,
+                        retry_for_unexpected: #retry_for_unexpected,
                         acks_late: #acks_late,
                     };