+++ src/app/mod.rs
@@ -59,8 +59,9 @@ where
                 task_options: TaskOptions {
                     timeout: None,
                     max_retries: None,
-                    min_retry_delay: Some(0),
-                    max_retry_delay: Some(3600),
+                    min_retry_delay: None,
+                    max_retry_delay: None,
+                    retry_for_unexpected: None,
                     acks_late: Some(false),
                 },
                 task_routes: vec![],
@@ -68,49 +69,67 @@ where
         }
     }
 
-    /// Set the name of the default queue.
+    /// Set the name of the default queue to something other than "celery".
     pub fn default_queue(mut self, queue_name: &str) -> Self {
         self.config.default_queue = queue_name.into();
         self
     }
 
-    /// Set the prefetch count.
+    /// Set the prefetch count. The default value depends on the broker implementation,
+    /// but it's recommended that you always set this to a value that works best
+    /// for your application.
+    ///
+    /// This may take some tuning, as it depends on a lot of factors, such
+    /// as whether your tasks are IO bound (higher prefetch count is better) or CPU bound (lower
+    /// prefetch count is better).
     pub fn prefetch_count(mut self, prefetch_count: u16) -> Self {
         self.config.broker_builder = self.config.broker_builder.prefetch_count(prefetch_count);
         self
     }
 
-    /// Set the broker heartbeat.
+    /// Set the broker heartbeat. The default value depends on the broker implementation.
     pub fn heartbeat(mut self, heartbeat: Option<u16>) -> Self {
         self.config.broker_builder = self.config.broker_builder.heartbeat(heartbeat);
         self
     }
 
-    /// Set a default timeout for tasks.
+    /// Set an app-level timeout for tasks (see
+    /// [`TaskOption::timeout`](task/struct.TaskOptions.html#structfield.timeout)).
     pub fn task_timeout(mut self, task_timeout: u32) -> Self {
         self.config.task_options.timeout = Some(task_timeout);
         self
     }
 
-    /// Set a default maximum number of retries for tasks.
+    /// Set an app-level maximum number of retries for tasks (see
+    /// [`TaskOption::max_retries`](task/struct.TaskOptions.html#structfield.max_retries)).
     pub fn task_max_retries(mut self, task_max_retries: u32) -> Self {
         self.config.task_options.max_retries = Some(task_max_retries);
         self
     }
 
-    /// Set a default minimum retry delay for tasks.
+    /// Set an app-level minimum retry delay for tasks (see
+    /// [`TaskOption::min_retry_delay`](task/struct.TaskOptions.html#structfield.min_retry_delay)).
     pub fn task_min_retry_delay(mut self, task_min_retry_delay: u32) -> Self {
         self.config.task_options.min_retry_delay = Some(task_min_retry_delay);
         self
     }
 
-    /// Set a default maximum retry delay for tasks.
+    /// Set an app-level maximum retry delay for tasks (see
+    /// [`TaskOption::max_retry_delay`](task/struct.TaskOptions.html#structfield.max_retry_delay)).
     pub fn task_max_retry_delay(mut self, task_max_retry_delay: u32) -> Self {
         self.config.task_options.max_retry_delay = Some(task_max_retry_delay);
         self
     }
 
-    /// Set whether by default a task is acknowledged before or after execution.
+    /// Set whether by default `UnexpectedError`s should be retried for (see
+    /// [`TaskOption::retry_for_unexpected`](task/struct.TaskOptions.html#structfield.retry_for_unexpected)).
+    pub fn task_retry_for_unexpected(mut self, retry_for_unexpected: bool) -> Self {
+        self.config.task_options.retry_for_unexpected = Some(retry_for_unexpected);
+        self
+    }
+
+    /// Set whether by default a task is acknowledged before or after execution (see
+    /// [`TaskOption::acks_late`](task/struct.TaskOptions.html#structfield.acks_late)).
     pub fn acks_late(mut self, acks_late: bool) -> Self {
         self.config.task_options.acks_late = Some(acks_late);
         self
@@ -345,7 +364,7 @@ where
         }
 
         // If acks_late is false, we acknowledge the message before tracing it.
-        if !tracer.get_task_options().acks_late.unwrap_or_default() {
+        if !tracer.acks_late() {
             self.broker
                 .ack(&delivery)
                 .await
@@ -367,7 +386,7 @@ where
         }
 
         // If we have not done it before, we have to acknowledge the message now.
-        if tracer.get_task_options().acks_late.unwrap_or_default() {
+        if tracer.acks_late() {
             self.broker
                 .ack(&delivery)
                 .await