+++ src/lib.rs
@@ -90,11 +90,12 @@ mod codegen;
 ///
 /// - `name`: The name to use when registering the task. Should be unique. If not given the name
 /// will be set to the name of the function being decorated.
-/// - `timeout`: Corresponds to [`Task::timeout`](trait.Task.html#method.timeout).
-/// - `max_retries`: Corresponds to [`Task::max_retries`](trait.Task.html#method.max_retries).
-/// - `min_retry_delay`: Corresponds to [`Task::min_retry_delay`](trait.Task.html#method.min_retry_delay).
-/// - `max_retry_delay`: Corresponds to [`Task::max_retry_delay`](trait.Task.html#method.max_retry_delay).
-/// - `acks_late`: Corresponds to [`Task::acks_late`](trait.Task.html#method.acks_late).
+/// - `timeout`: Set a task-level [`TaskOptions::timeout`](task/struct.TaskOptions.html#structfield.timeout).
+/// - `max_retries`: Set a task-level [`TaskOptions::max_retries`](task/struct.TaskOptions.html#structfield.max_retries).
+/// - `min_retry_delay`: Set a task-level [`TaskOptions::min_retry_delay`](task/struct.TaskOptions.html#structfield.min_retry_delay).
+/// - `max_retry_delay`: Set a task-level [`TaskOptions::max_retry_delay`](task/struct.TaskOptions.html#structfield.max_retry_delay).
+/// - `retry_for_unexpected`: Set a task-level [`TaskOptions::retry_for_unexpected`](task/struct.TaskOptions.html#structfield.retry_for_unexpected).
+/// - `acks_late`: Set a task-level [`TaskOptions::acks_late`](task/struct.TaskOptions.html#structfield.acks_late).
 /// - `bind`: A bool. If true, the task will be run like an instance method and so the function's
 /// first argument should be a reference to `Self`. Note however that Rust won't allow you to call
 /// the argument `self`. Instead, you could use `task` or just `t`.